requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
langchain==0.0.335
langchain-ollama==0.0.1
langchain-community==0.0.1
neo4j==5.14.1
python-dotenv==1.0.0
PyPDF2==3.0.1
pydantic==2.4.2